# docs/openapi.yaml
openapi: 3.1.0
info:
 title: Public User Forum
 description: Specification for all of the endpoints on an api used for a web forum. Also included is user accounts
 version: 1.0.0
servers:
 - url: http://example.com/api
   description: placeholder URL will need to replace
...
paths:
...
 /forums/{forumID}/threads/{threadID}:
  get:
    tags:
      - Threads
    summary: Get thread by ID
    description: Get thread by ID
    operationId: getThread
    parameters:
      - $ref: '#/components/parameters/forumID'
      - $ref: '#/components/parameters/threadID'
    responses:
      '200':
        description: Sucessful operation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/threadDetails'
      default:
        description: Unable to return values
  patch:
    tags:
      - Threads
    summary: Update thread
    description: Update thread
    operationId: updateThread
    parameters:
      - $ref: '#/components/parameters/forumID'
      - $ref: '#/components/parameters/threadID'
    requestBody: 
      required: true
      description: update thread
      content:
        application/json:
          schema:
            type: object
            properties:
              threadPost:
                type: string
                examples: [Updated post is here now]
    responses:
      '200':
        description: Sucessful operation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/threadDetails'
      default:
        description: Unable to return values
  delete:
    tags:
      - Threads
    summary: delete thread
    description: delete thread
    operationId: deleteThread
    parameters:
      - $ref: '#/components/parameters/forumID'
      - $ref: '#/components/parameters/threadID'
    responses:
      '200':
        description: Thread sucessfully deleted
...
components:
  parameters:
...
    forumID:
      name: forumID
      in: path
      required: true
      schema:
        type: integer
      description: ID of the forum
...
    threadID:
      name: threadID
      in: path
      required: true
      schema:
        type: integer
      description: ID of the thread
...
  schemas:
...
    threadDetails:
      type: object
      required:
        - threadTitle
        - threadPost
        - creatorID
      properties:
        threadID:
          type: integer
          examples: [1]
        createdAt:
          type: string
          examples: [01/01/26]
        threadTitle:
          type: string
          examples: [Why are taco's so good?]
        threadPost:
          type: string
          examples: [The flavors of the tacos just work so well. Does anyone know the reason for this?]
        creatorID:
          type: integer
          examples: [1]
...